fix merge changes for most part
