+++ kernel/src/kernel.rs
@@ -11,7 +11,6 @@
 use core::cell::Cell;
 use core::ptr::NonNull;
 
-use crate::allocator::IntoCortexMPU;
 use crate::capabilities;
 use crate::config;
 use crate::debug;
@@ -21,6 +20,7 @@ use crate::grant::{AllowRoSize, AllowRwSize, Grant, UpcallSize};
 use crate::ipc;
 use crate::memop;
 use crate::platform::chip::Chip;
+use crate::platform::mpu::MPU;
 use crate::platform::platform::ContextSwitchCallback;
 use crate::platform::platform::KernelResources;
 use crate::platform::platform::{ProcessFault, SyscallDriverLookup, SyscallFilter};
@@ -551,19 +551,13 @@ impl Kernel {
                         .context_switch_callback()
                         .context_switch_hook(process);
                     let mpu_configured_capability = process.setup_mpu();
-                    let mpu_enabled_capability = match chip.mpu().into_cortex_mpu() {
-                        crate::allocator::CortexMpuTypes::Sixteen(mpu) => mpu.enable_app_mpu(),
-                        crate::allocator::CortexMpuTypes::Eight(mpu) => mpu.enable_app_mpu(),
-                    };
+                    let mpu_enabled_capability = chip.mpu().enable_app_mpu();
+
                     scheduler_timer.arm();
                     let context_switch_reason =
                         process.switch_to(mpu_configured_capability, mpu_enabled_capability);
                     scheduler_timer.disarm();
-                    match chip.mpu().into_cortex_mpu() {
-                        crate::allocator::CortexMpuTypes::Sixteen(mpu) => mpu.disable_app_mpu(),
-                        crate::allocator::CortexMpuTypes::Eight(mpu) => mpu.disable_app_mpu(),
-                    };
-
+                    chip.mpu().disable_app_mpu();
                     // Now the process has returned back to the kernel. Check
                     // why and handle the process as appropriate.
                     match context_switch_reason {
